% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sender.R
\name{check_credentials}
\alias{check_credentials}
\title{Credentials check}
\usage{
check_credentials(group, from, key, verbose = TRUE)
}
\arguments{
\item{group}{An \code{integer} representing your group number (should be between 1 and 17).}

\item{from}{A \code{string} representing your group email address in the correct format.}

\item{key}{A \code{string} representing the unique key assigned to your group.}

\item{verbose}{A \code{logical} value to indicate if an affirmative output should be printed (default \code{TRUE}).}
}
\description{
This function performs checks on your credentials. If there are no errors
this function returns \code{TRUE}.
}
