% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sender.R
\name{check_prediction}
\alias{check_prediction}
\title{Prediction object check}
\usage{
check_prediction(prediction, verbose = TRUE)
}
\arguments{
\item{prediction}{A \code{list} containing 7 elements representing your forecasts (refer to project description on Piazza).}

\item{verbose}{A \code{logical} value to indicate if an affirmative output should be printed (default \code{TRUE}).}
}
\description{
This function performs checks on the format of the prediction object. If there are no errors
this function returns \code{TRUE}.
}
