% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sender.R
\name{send_prediction}
\alias{send_prediction}
\title{Submit forecast objects for STAT 463}
\usage{
send_prediction(group, prediction, from, key)
}
\arguments{
\item{group}{An \code{integer} representing your group number (should be between 1 and 17).}

\item{prediction}{A \code{list} containing 7 elements representing your forecasts (refer to project description on Piazza).}

\item{from}{A \code{string} representing your group email address in the correct format.}

\item{key}{A \code{string} representing the unique key assigned to your group.}
}
\description{
This function performs check on the format of submission and subsequently submits the forecast object.
}
